LayoutTests/fast/selectors/matches-backtracking-expected.txt

Check backtracking with :matches(selectorList).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing "testcase:matches(bb > aa > aa > bb testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(bb > aa > aa > bb testcase)').length is 4
PASS document.querySelectorAll('#test-root testcase:matches(bb > aa > aa > bb testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(bb > aa > aa > bb testcase)')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:matches(bb > aa > aa > bb testcase)')[2].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:matches(bb > aa > aa > bb testcase)')[3].id is "testcase4"
PASS styledElements is ["testcase1","testcase2","testcase3","testcase4"]

Testing "testcase:matches(bb > aa > aa > bb > aa testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(bb > aa > aa > bb > aa testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(bb > aa > aa > bb > aa testcase)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:matches(bb > aa > aa > bb > aa testcase)')[1].id is "testcase4"
PASS styledElements is ["testcase3","testcase4"]

Testing "testcase:matches(bb > * > aa > bb > aa testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(bb > * > aa > bb > aa testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(bb > * > aa > bb > aa testcase)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:matches(bb > * > aa > bb > aa testcase)')[1].id is "testcase4"
PASS styledElements is ["testcase3","testcase4"]

Testing "testcase:matches(bb > aa > * > bb > aa testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(bb > aa > * > bb > aa testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(bb > aa > * > bb > aa testcase)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:matches(bb > aa > * > bb > aa testcase)')[1].id is "testcase4"
PASS styledElements is ["testcase3","testcase4"]

Testing "testcase:matches(bb > * > * > bb > aa testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(bb > * > * > bb > aa testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(bb > * > * > bb > aa testcase)')[0].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:matches(bb > * > * > bb > aa testcase)')[1].id is "testcase4"
PASS styledElements is ["testcase3","testcase4"]

Testing "testcase:matches(* > * > * > bb > aa testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(* > * > * > bb > aa testcase)').length is 4
PASS document.querySelectorAll('#test-root testcase:matches(* > * > * > bb > aa testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(* > * > * > bb > aa testcase)')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:matches(* > * > * > bb > aa testcase)')[2].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:matches(* > * > * > bb > aa testcase)')[3].id is "testcase4"
PASS styledElements is ["testcase1","testcase2","testcase3","testcase4"]

Testing "testcase:matches(aa > aa > aa > bb testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(aa > aa > aa > bb testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(aa > aa > aa > bb testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(aa > aa > aa > bb testcase)')[1].id is "testcase2"
PASS styledElements is ["testcase1","testcase2"]

Testing "testcase:matches(aa > aa > * > bb testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(aa > aa > * > bb testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(aa > aa > * > bb testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(aa > aa > * > bb testcase)')[1].id is "testcase2"
PASS styledElements is ["testcase1","testcase2"]

Testing "testcase:matches(aa > * > * > bb testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(aa > * > * > bb testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(aa > * > * > bb testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(aa > * > * > bb testcase)')[1].id is "testcase2"
PASS styledElements is ["testcase1","testcase2"]

Testing "testcase:matches(aa > aa > aa > * testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(aa > aa > aa > * testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(aa > aa > aa > * testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(aa > aa > aa > * testcase)')[1].id is "testcase2"
PASS styledElements is ["testcase1","testcase2"]

Testing "testcase:matches(aa > * > * > * testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(aa > * > * > * testcase)').length is 4
PASS document.querySelectorAll('#test-root testcase:matches(aa > * > * > * testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(aa > * > * > * testcase)')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:matches(aa > * > * > * testcase)')[2].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:matches(aa > * > * > * testcase)')[3].id is "testcase4"
PASS styledElements is ["testcase1","testcase2","testcase3","testcase4"]

Testing "testcase:matches(* > * > * > bb testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(* > * > * > bb testcase)').length is 4
PASS document.querySelectorAll('#test-root testcase:matches(* > * > * > bb testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(* > * > * > bb testcase)')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:matches(* > * > * > bb testcase)')[2].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:matches(* > * > * > bb testcase)')[3].id is "testcase4"
PASS styledElements is ["testcase1","testcase2","testcase3","testcase4"]

Testing "testcase:matches(* > * > * > * testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(* > * > * > * testcase)').length is 4
PASS document.querySelectorAll('#test-root testcase:matches(* > * > * > * testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(* > * > * > * testcase)')[1].id is "testcase2"
PASS document.querySelectorAll('#test-root testcase:matches(* > * > * > * testcase)')[2].id is "testcase3"
PASS document.querySelectorAll('#test-root testcase:matches(* > * > * > * testcase)')[3].id is "testcase4"
PASS styledElements is ["testcase1","testcase2","testcase3","testcase4"]

Testing "testcase:matches(dd + dd + cc + dd ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(dd + dd + cc + dd ~ testcase)').length is 1
PASS document.querySelectorAll('#test-root testcase:matches(dd + dd + cc + dd ~ testcase)')[0].id is "testcase3"
PASS styledElements is ["testcase3"]

Testing "testcase:matches(dd + dd + cc + * ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(dd + dd + cc + * ~ testcase)').length is 1
PASS document.querySelectorAll('#test-root testcase:matches(dd + dd + cc + * ~ testcase)')[0].id is "testcase3"
PASS styledElements is ["testcase3"]

Testing "testcase:matches(dd + dd + * + dd ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(dd + dd + * + dd ~ testcase)').length is 1
PASS document.querySelectorAll('#test-root testcase:matches(dd + dd + * + dd ~ testcase)')[0].id is "testcase3"
PASS styledElements is ["testcase3"]

Testing "testcase:matches(dd + dd + * + * ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(dd + dd + * + * ~ testcase)').length is 1
PASS document.querySelectorAll('#test-root testcase:matches(dd + dd + * + * ~ testcase)')[0].id is "testcase3"
PASS styledElements is ["testcase3"]

Testing "testcase:matches(dd + * + cc + dd ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(dd + * + cc + dd ~ testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(dd + * + cc + dd ~ testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(dd + * + cc + dd ~ testcase)')[1].id is "testcase3"
PASS styledElements is ["testcase1","testcase3"]

Testing "testcase:matches(* + dd + cc + dd ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(* + dd + cc + dd ~ testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(* + dd + cc + dd ~ testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(* + dd + cc + dd ~ testcase)')[1].id is "testcase3"
PASS styledElements is ["testcase1","testcase3"]

Testing "testcase:matches(* + * + cc + dd ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(* + * + cc + dd ~ testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(* + * + cc + dd ~ testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(* + * + cc + dd ~ testcase)')[1].id is "testcase3"
PASS styledElements is ["testcase1","testcase3"]

Testing "testcase:matches(cc + cc + cc + cc + dd ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(cc + cc + cc + cc + dd ~ testcase)').length is 1
PASS document.querySelectorAll('#test-root testcase:matches(cc + cc + cc + cc + dd ~ testcase)')[0].id is "testcase1"
PASS styledElements is ["testcase1"]

Testing "testcase:matches(cc + cc + cc + cc + * ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(cc + cc + cc + cc + * ~ testcase)').length is 1
PASS document.querySelectorAll('#test-root testcase:matches(cc + cc + cc + cc + * ~ testcase)')[0].id is "testcase1"
PASS styledElements is ["testcase1"]

Testing "testcase:matches(cc + cc + cc + * + dd ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(cc + cc + cc + * + dd ~ testcase)').length is 1
PASS document.querySelectorAll('#test-root testcase:matches(cc + cc + cc + * + dd ~ testcase)')[0].id is "testcase1"
PASS styledElements is ["testcase1"]

Testing "testcase:matches(cc + cc + cc + * + * ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(cc + cc + cc + * + * ~ testcase)').length is 1
PASS document.querySelectorAll('#test-root testcase:matches(cc + cc + cc + * + * ~ testcase)')[0].id is "testcase1"
PASS styledElements is ["testcase1"]

Testing "testcase:matches(cc + cc + * + * + * ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(cc + cc + * + * + * ~ testcase)').length is 1
PASS document.querySelectorAll('#test-root testcase:matches(cc + cc + * + * + * ~ testcase)')[0].id is "testcase1"
PASS styledElements is ["testcase1"]

Testing "testcase:matches(cc + * + * + * + * ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(cc + * + * + * + * ~ testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(cc + * + * + * + * ~ testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(cc + * + * + * + * ~ testcase)')[1].id is "testcase3"
PASS styledElements is ["testcase1","testcase3"]

Testing "testcase:matches(* + cc + * + * + * ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(* + cc + * + * + * ~ testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(* + cc + * + * + * ~ testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(* + cc + * + * + * ~ testcase)')[1].id is "testcase3"
PASS styledElements is ["testcase1","testcase3"]

Testing "testcase:matches(* + * + * + * + * ~ testcase)"
PASS document.querySelectorAll('#test-root testcase:matches(* + * + * + * + * ~ testcase)').length is 2
PASS document.querySelectorAll('#test-root testcase:matches(* + * + * + * + * ~ testcase)')[0].id is "testcase1"
PASS document.querySelectorAll('#test-root testcase:matches(* + * + * + * + * ~ testcase)')[1].id is "testcase3"
PASS styledElements is ["testcase1","testcase3"]

Testing ":matches(ee) > ee > ee > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(ee) > ee > ee > ee > ff testcase').length is 0
PASS styledElements is []

Testing ":matches(testcase) > ee > ee > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(testcase) > ee > ee > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :matches(testcase) > ee > ee > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":matches(ee, foo, bar, baz) > ee > ee > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(ee, foo, bar, baz) > ee > ee > ee > ff testcase').length is 0
PASS styledElements is []

Testing ":matches(foo, bar, baz, testcase) > ee > ee > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(foo, bar, baz, testcase) > ee > ee > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :matches(foo, bar, baz, testcase) > ee > ee > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":matches(ee) > ee > * > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(ee) > ee > * > ee > ff testcase').length is 0
PASS styledElements is []

Testing ":matches(testcase) > ee > * > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(testcase) > ee > * > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :matches(testcase) > ee > * > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":matches(ee, foo, bar, baz) > ee > * > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(ee, foo, bar, baz) > ee > * > ee > ff testcase').length is 0
PASS styledElements is []

Testing ":matches(foo, bar, baz, testcase) > ee > * > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(foo, bar, baz, testcase) > ee > * > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :matches(foo, bar, baz, testcase) > ee > * > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":matches(ee) > ee > ee > * > ff testcase"
PASS document.querySelectorAll('#test-root :matches(ee) > ee > ee > * > ff testcase').length is 0
PASS styledElements is []

Testing ":matches(testcase) > ee > ee > * > ff testcase"
PASS document.querySelectorAll('#test-root :matches(testcase) > ee > ee > * > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :matches(testcase) > ee > ee > * > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":matches(ee, foo, bar, baz) > ee > ee > * > ff testcase"
PASS document.querySelectorAll('#test-root :matches(ee, foo, bar, baz) > ee > ee > * > ff testcase').length is 0
PASS styledElements is []

Testing ":matches(foo, bar, baz, testcase) > ee > ee > * > ff testcase"
PASS document.querySelectorAll('#test-root :matches(foo, bar, baz, testcase) > ee > ee > * > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :matches(foo, bar, baz, testcase) > ee > ee > * > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":matches(ee) > * > ee > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(ee) > * > ee > ee > ff testcase').length is 2
PASS document.querySelectorAll('#test-root :matches(ee) > * > ee > ee > ff testcase')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :matches(ee) > * > ee > ee > ff testcase')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing ":matches(testcase) > * > ee > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(testcase) > * > ee > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :matches(testcase) > * > ee > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":matches(ee, foo, bar, baz) > * > ee > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(ee, foo, bar, baz) > * > ee > ee > ff testcase').length is 2
PASS document.querySelectorAll('#test-root :matches(ee, foo, bar, baz) > * > ee > ee > ff testcase')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :matches(ee, foo, bar, baz) > * > ee > ee > ff testcase')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing ":matches(foo, bar, baz, testcase) > * > ee > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(foo, bar, baz, testcase) > * > ee > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :matches(foo, bar, baz, testcase) > * > ee > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":matches(ee) > * > * > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(ee) > * > * > ee > ff testcase').length is 2
PASS document.querySelectorAll('#test-root :matches(ee) > * > * > ee > ff testcase')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :matches(ee) > * > * > ee > ff testcase')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing ":matches(testcase) > * > * > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(testcase) > * > * > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :matches(testcase) > * > * > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":matches(ee, foo, bar, baz) > * > * > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(ee, foo, bar, baz) > * > * > ee > ff testcase').length is 2
PASS document.querySelectorAll('#test-root :matches(ee, foo, bar, baz) > * > * > ee > ff testcase')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :matches(ee, foo, bar, baz) > * > * > ee > ff testcase')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing ":matches(foo, bar, baz, testcase) > * > * > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(foo, bar, baz, testcase) > * > * > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :matches(foo, bar, baz, testcase) > * > * > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":matches(ee) > * > * > ee > * testcase"
PASS document.querySelectorAll('#test-root :matches(ee) > * > * > ee > * testcase').length is 2
PASS document.querySelectorAll('#test-root :matches(ee) > * > * > ee > * testcase')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :matches(ee) > * > * > ee > * testcase')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing ":matches(testcase) > * > * > ee > * testcase"
PASS document.querySelectorAll('#test-root :matches(testcase) > * > * > ee > * testcase').length is 1
PASS document.querySelectorAll('#test-root :matches(testcase) > * > * > ee > * testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":matches(ee, foo, bar, baz) > * > * > ee > * testcase"
PASS document.querySelectorAll('#test-root :matches(ee, foo, bar, baz) > * > * > ee > * testcase').length is 2
PASS document.querySelectorAll('#test-root :matches(ee, foo, bar, baz) > * > * > ee > * testcase')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :matches(ee, foo, bar, baz) > * > * > ee > * testcase')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing ":matches(foo, bar, baz, testcase) > * > * > ee > * testcase"
PASS document.querySelectorAll('#test-root :matches(foo, bar, baz, testcase) > * > * > ee > * testcase').length is 1
PASS document.querySelectorAll('#test-root :matches(foo, bar, baz, testcase) > * > * > ee > * testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":matches(ee) > * > * > * > * testcase"
PASS document.querySelectorAll('#test-root :matches(ee) > * > * > * > * testcase').length is 2
PASS document.querySelectorAll('#test-root :matches(ee) > * > * > * > * testcase')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :matches(ee) > * > * > * > * testcase')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing ":matches(testcase) > * > * > * > * testcase"
PASS document.querySelectorAll('#test-root :matches(testcase) > * > * > * > * testcase').length is 2
PASS document.querySelectorAll('#test-root :matches(testcase) > * > * > * > * testcase')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :matches(testcase) > * > * > * > * testcase')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing ":matches(ee, foo, bar, baz) > * > * > * > * testcase"
PASS document.querySelectorAll('#test-root :matches(ee, foo, bar, baz) > * > * > * > * testcase').length is 2
PASS document.querySelectorAll('#test-root :matches(ee, foo, bar, baz) > * > * > * > * testcase')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :matches(ee, foo, bar, baz) > * > * > * > * testcase')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing ":matches(foo, bar, baz, testcase) > * > * > * > * testcase"
PASS document.querySelectorAll('#test-root :matches(foo, bar, baz, testcase) > * > * > * > * testcase').length is 2
PASS document.querySelectorAll('#test-root :matches(foo, bar, baz, testcase) > * > * > * > * testcase')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :matches(foo, bar, baz, testcase) > * > * > * > * testcase')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing ":matches(bb > * > * > bb > aa testcase, foo, bar, baz) > ee > ee > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(bb > * > * > bb > aa testcase, foo, bar, baz) > ee > ee > ee > ff testcase').length is 0
PASS styledElements is []

Testing ":matches(aa > * > * > bb > aa testcase, foo, bar, baz) > ee > ee > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(aa > * > * > bb > aa testcase, foo, bar, baz) > ee > ee > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :matches(aa > * > * > bb > aa testcase, foo, bar, baz) > ee > ee > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":matches(bb > * > * > bb > aa testcase, foo, bar, baz) > ee > * > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(bb > * > * > bb > aa testcase, foo, bar, baz) > ee > * > ee > ff testcase').length is 0
PASS styledElements is []

Testing ":matches(aa > * > * > bb > aa testcase, foo, bar, baz) > ee > * > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(aa > * > * > bb > aa testcase, foo, bar, baz) > ee > * > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :matches(aa > * > * > bb > aa testcase, foo, bar, baz) > ee > * > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":matches(bb > * > * > bb > aa testcase, foo, bar, baz) > ee > ee > * > ff testcase"
PASS document.querySelectorAll('#test-root :matches(bb > * > * > bb > aa testcase, foo, bar, baz) > ee > ee > * > ff testcase').length is 0
PASS styledElements is []

Testing ":matches(aa > * > * > bb > aa testcase, foo, bar, baz) > ee > ee > * > ff testcase"
PASS document.querySelectorAll('#test-root :matches(aa > * > * > bb > aa testcase, foo, bar, baz) > ee > ee > * > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :matches(aa > * > * > bb > aa testcase, foo, bar, baz) > ee > ee > * > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":matches(bb > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > ee > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(bb > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > ee > ee > ff testcase').length is 0
PASS styledElements is []

Testing ":matches(aa > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > ee > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(aa > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > ee > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :matches(aa > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > ee > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":matches(bb > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > * > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(bb > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > * > ee > ff testcase').length is 0
PASS styledElements is []

Testing ":matches(aa > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > * > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(aa > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > * > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :matches(aa > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > * > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":matches(bb > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > ee > * > ff testcase"
PASS document.querySelectorAll('#test-root :matches(bb > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > ee > * > ff testcase').length is 0
PASS styledElements is []

Testing ":matches(aa > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > ee > * > ff testcase"
PASS document.querySelectorAll('#test-root :matches(aa > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > ee > * > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :matches(aa > * > * > bb > aa * + * + cc + dd ~ testcase, foo, bar, baz) > ee > ee > * > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":matches(bb > * > * > bb > aa dd + dd + * + * ~ testcase, foo, bar, baz) > ee > ee > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(bb > * > * > bb > aa dd + dd + * + * ~ testcase, foo, bar, baz) > ee > ee > ee > ff testcase').length is 0
PASS styledElements is []

Testing ":matches(aa > * > * > bb > aa cc + dd + * + * ~ testcase, foo, bar, baz) > ee > ee > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(aa > * > * > bb > aa cc + dd + * + * ~ testcase, foo, bar, baz) > ee > ee > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :matches(aa > * > * > bb > aa cc + dd + * + * ~ testcase, foo, bar, baz) > ee > ee > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":matches(bb > * > * > bb > aa dd + dd + * + * ~ testcase, foo, bar, baz) > ee > * > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(bb > * > * > bb > aa dd + dd + * + * ~ testcase, foo, bar, baz) > ee > * > ee > ff testcase').length is 0
PASS styledElements is []

Testing ":matches(aa > * > * > bb > aa cc + dd + * + * ~ testcase, foo, bar, baz) > ee > * > ee > ff testcase"
PASS document.querySelectorAll('#test-root :matches(aa > * > * > bb > aa cc + dd + * + * ~ testcase, foo, bar, baz) > ee > * > ee > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :matches(aa > * > * > bb > aa cc + dd + * + * ~ testcase, foo, bar, baz) > ee > * > ee > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":matches(bb > * > * > bb > aa dd + dd + * + * ~ testcase, foo, bar, baz) > ee > ee > * > ff testcase"
PASS document.querySelectorAll('#test-root :matches(bb > * > * > bb > aa dd + dd + * + * ~ testcase, foo, bar, baz) > ee > ee > * > ff testcase').length is 0
PASS styledElements is []

Testing ":matches(aa > * > * > bb > aa cc + dd + * + * ~ testcase, foo, bar, baz) > ee > ee > * > ff testcase"
PASS document.querySelectorAll('#test-root :matches(aa > * > * > bb > aa cc + dd + * + * ~ testcase, foo, bar, baz) > ee > ee > * > ff testcase').length is 1
PASS document.querySelectorAll('#test-root :matches(aa > * > * > bb > aa cc + dd + * + * ~ testcase, foo, bar, baz) > ee > ee > * > ff testcase')[0].id is "testcase2"
PASS styledElements is ["testcase2"]

Testing ":matches(* > * > * > * testcase, foo, bar, baz) > * > * > * > * testcase"
PASS document.querySelectorAll('#test-root :matches(* > * > * > * testcase, foo, bar, baz) > * > * > * > * testcase').length is 2
PASS document.querySelectorAll('#test-root :matches(* > * > * > * testcase, foo, bar, baz) > * > * > * > * testcase')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :matches(* > * > * > * testcase, foo, bar, baz) > * > * > * > * testcase')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing ":matches(* > * > * > bb testcase, foo, bar, baz) > * > * > * > * testcase"
PASS document.querySelectorAll('#test-root :matches(* > * > * > bb testcase, foo, bar, baz) > * > * > * > * testcase').length is 2
PASS document.querySelectorAll('#test-root :matches(* > * > * > bb testcase, foo, bar, baz) > * > * > * > * testcase')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root :matches(* > * > * > bb testcase, foo, bar, baz) > * > * > * > * testcase')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

Testing "bb > :matches(* > * > * > * testcase, foo, bar, baz) > * > * > * > * testcase"
PASS document.querySelectorAll('#test-root bb > :matches(* > * > * > * testcase, foo, bar, baz) > * > * > * > * testcase').length is 2
PASS document.querySelectorAll('#test-root bb > :matches(* > * > * > * testcase, foo, bar, baz) > * > * > * > * testcase')[0].id is "testcase2"
PASS document.querySelectorAll('#test-root bb > :matches(* > * > * > * testcase, foo, bar, baz) > * > * > * > * testcase')[1].id is "testcase4"
PASS styledElements is ["testcase2","testcase4"]

PASS successfullyParsed is true

TEST COMPLETE
